refactor: Provider comp inside of UserContext.js
