+++ client/src/App.js
@@ -1,4 +1,4 @@
-import React, { useState, useMemo } from "react";
+import React from "react";
 import { MuiThemeProvider } from "@material-ui/core";
 import { BrowserRouter, Route, Switch } from "react-router-dom";
 import Login from "./pages/Login.jsx";
@@ -6,21 +6,14 @@ import SignUp from "./pages/SignUp.jsx";
 import { theme } from "./themes/theme";
 import LandingPage from "./pages/Landing.jsx";
 import Board from "./pages/Board";
-import { UserContext } from "./contexts/UserContext";
+import { UserContextProvider } from "./contexts/UserContext";
 import { ProtectedRoute } from "./auth/ProtectedRoute";
 
 import "./App.css";
 
 function App() {
-  const [loggedIn, setLoggedIn] = useState(false);
-  const [user, setUser] = useState({});
-  const providerValue = useMemo(
-    () => ({ user, setUser, loggedIn, setLoggedIn }),
-    [user, setUser, loggedIn, setLoggedIn]
-  );
-
   return (
-    <UserContext.Provider value={providerValue}>
+    <UserContextProvider>
       <MuiThemeProvider theme={theme}>
         <BrowserRouter>
           <Switch>
@@ -32,7 +25,7 @@ function App() {
           </Switch>
         </BrowserRouter>
       </MuiThemeProvider>
-    </UserContext.Provider>
+    </UserContextProvider>
   );
 }
 export default App;
